binfmt_elf: Update READ_IMPLIES_EXEC logic for modern CPUs
The READ_IMPLIES_EXEC work-around was designed for old CPUs lacking NX (to
have the visible permission flags on memory regions reflect reality: they
are all executable), and for old toolchains that lacked the ELF
PT_GNU_STACK marking (under the assumption than toolchains that couldn't
even specify memory protection flags may have it wrong for all memory
regions).

This logic is sensible, but was implemented in a way that equated having a
PT_GNU_STACK marked executable as being as "broken" as lacking the
PT_GNU_STACK marking entirely.  This is not a reasonable assumption for
CPUs that have had NX support from the start (or very close to the start).
This confusion has led to situations where modern 64-bit programs with
explicitly marked executable stack are forced into the READ_IMPLIES_EXEC
state when no such thing is needed.  (And leads to unexpected failures
when mmap()ing regions of device driver memory that wish to disallow
VM_EXEC[1].)

To fix this, elf_read_implies_exec() is adjusted on arm64 (where NX has
always existed and all toolchains include PT_GNU_STACK), and x86 is
adjusted to handle this combination of possible outcomes:

              CPU: | lacks NX  | has NX, ia32     | has NX, x86_64   |
 ELF:              |           |                  |                  |
 ------------------------------|------------------|------------------|
 missing GNU_STACK | needs RIE | needs RIE        | no RIE           |
 GNU_STACK == RWX  | needs RIE | no RIE: stack X  | no RIE: stack X  |
 GNU_STACK == RW   | needs RIE | no RIE: stack NX | no RIE: stack NX |

This has the effect of making binfmt_elf's EXSTACK_DEFAULT actually take
on the correct architecture default of being non-executable on arm64 and
x86_64, and being executable on ia32.

[1] https://lkml.kernel.org/r/20190418055759.GA3155@mellanox.com

Link: http://lkml.kernel.org/r/20190423181210.GA2443@beast
Signed-off-by: Kees Cook <keescook@chromium.org>
Suggested-by: Hector Marco-Gisbert <hecmargi@upv.es>
Cc: Jason Gunthorpe <jgg@mellanox.com>
Cc: Thomas Gleixner <tglx@linutronix.de>
Cc: Andy Lutomirski <luto@amacapital.net>
Cc: Kernel Hardening <kernel-hardening@lists.openwall.com>
Cc: Mark Rutland <mark.rutland@arm.com>
Cc: Ingo Molnar <mingo@elte.hu>
Cc: "H. Peter Anvin" <hpa@zytor.com>
Cc: Peter Zijlstra <a.p.zijlstra@chello.nl>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>