signal: adjust error codes according to restore_user_sigmask()
A regression caused by 854a6ed56839 ("signal: Add restore_user_sigmask()")
caused users of epoll_pwait, io_pgetevents, and ppoll to notice a latent
problem in signal handling during these syscalls.

That patch (854a6ed56839) moved the signal_pending() check closer to
restoring of the user sigmask.  But, it failed to update the error code
accordingly.  From the userspace perspective, the patch increased the time
window for the signal discovery and subsequent delivery to the userspace,
but did not always adjust the errno afterwards.  The behavior before
854a6ed56839a was that the signals were dropped after the error code was
decided.  This resulted in lost signals but the userspace did not notice
it as the syscalls had finished executing the core functionality and the
error codes returned notified success.

For all the syscalls that receive a sigmask from the userland, the user
sigmask is to be in effect through the syscall execution.  At the end of
syscall, sigmask of the current process is restored to what it was before
the switch over to user sigmask.  But, for this to be true in practice,
the sigmask should be restored only at the the point we change the
saved_sigmask.  Anything before that loses signals.  And, anything after
is just pointless as the signal is already lost by restoring the sigmask.

Detailed issue discussion permalink:
https://lore.kernel.org/linux-fsdevel/20190427093319.sgicqik2oqkez3wk@dcvr/

Note that this patch returns interrupted errors (EINTR, ERESTARTNOHAND,
etc) only when there is no other error.  If there is a signal and an error
like EINVAL, the syscalls return -EINVAL rather than the interrupted error
codes.

Link: http://lkml.kernel.org/r/20190522032144.10995-1-deepa.kernel@gmail.com
Fixes: 854a6ed56839a40f ("signal: Add restore_user_sigmask()")
Signed-off-by: Deepa Dinamani <deepa.kernel@gmail.com>
Reported-by: Eric Wong <e@80x24.org>
Cc: Al Viro <viro@zeniv.linux.org.uk>
Cc: Arnd Bergmann <arnd@arndb.de>
Cc: Jens Axboe <axboe@kernel.dk>
Cc: Davidlohr Bueso <dave@stgolabs.net>
Cc: Jason Baron <jbaron@akamai.com>
Cc: Thomas Gleixner <tglx@linutronix.de>
Cc: Oleg Nesterov <oleg@redhat.com>
Cc: <stable@vger.kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>