mm/vmalloc: hugepage vmalloc mappings
For platforms that define HAVE_ARCH_HUGE_VMAP, have vmalloc try to
allocate and map PMD-level huge pages when size is appropriate, and
fallback if unsuccessful.

Using 2MB page mappings in an 8MB dentry cache hash brings dTLB misses for
linux kernel tree `git diff` from 45,000 to 8,000 on a Kaby Lake CPU wth
mitigations=off (performance is in the noise, under 1% difference, page
tables are likely to be well cached for this workload).

Link: http://lkml.kernel.org/r/20190528120453.27374-4-npiggin@gmail.com
Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
Cc: Mike Rapoport <rppt@linux.ibm.com>
Cc: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Cc: Toshi Kani <toshi.kani@hp.com>
Cc: Uladzislau Rezki <urezki@gmail.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>