mm/vmscan.c: add warn on inadvertently reclaiming mapped page
In isolate_lru_pages(), we check scan_control::may_unmap and set isolation
mode accordingly in order to not isolate from the lru list any page that
does not match the isolation mode.  For example, we should skip all sill
mapped pages if isolation mode is set to be ISOLATE_UNMAPPED.

So complain, while scanning the isolated pages, about the very unlikely
event that we hit a mapped page that we should never have isolated.  Note
no change is added in the current scanning behavior without VM debug
configured.

And cut off one line of comment that goes stale.

Link: http://lkml.kernel.org/r/20190526062353.14684-1-hdanton@sina.com
Signed-off-by: Hillf Danton <hdanton@sina.com>
Reviewed-by: Yang Shi <yang.shi@linux.alibaba.com>
Cc: Mel Gorman <mgorman@techsingularity.net>
Cc: Johannes Weiner <hannes@cmpxchg.org>
Cc: Minchan Kim <minchan@kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>